LayoutTests/fast/canvas/offscreen-giant-transfer-to-imagebitmap-expected.txt

CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 268435456).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 268435456).
